feat(tutorial): Update/cleanup tutorial
