--- examples/example.html
@@ -1,180 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <title>uPort Example</title>
-    <meta name="description" content="">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-    <script type="text/javascript" src="../dist/uport-connect.js"></script>
-    <style>
-      main { padding: 10px 30px; }
-      header { background: #EEE; }
-      .btn-sm{ padding: 3px 10px;
-               border-radius: 0; }
-    </style>
-  </head>
-  <body>
-    <main>
-      <header><h1 id="dappName">Uport Test</h1></header>
-      <hr/>
-      <section>
-        <h2>Connect uPort</h2>
-        <table>
-          <tobdy>
-            <tr><td><span>Address: </span><span id="addr"></span></td></tr>
-            <tr><td><span>Name: </span><span id="name"></span></td></tr>
-          </tobdy>
-        <table>
-        <button class="btn btn-sm btn-primary" id="connectUportBtn" onclick="loginBtnClick()">Connect uPort</button>
-      </section>
-      <hr/>
-      <section>
-        <h2>Set Status</h2>
-        <table>
-          <thead>
-            <tr>
-              <td>Show:</td>
-              <td></td>
-            </tr>
-          </thead>
-          <tobdy>
-            <tr>
-              <td><input type="text" id="sts" size="20" placeholder="Feeling good"/></td>
-              <td><button class="btn btn-sm btn-warning" onclick="updateStatus()">Update Status</button></td>
-            </tr>
-            <tr>
-              <td><span>Transaction Hash(ID): </span><span id="txHashStatus"></span></td>
-            </tr>
-            <tr>
-              <td><span>Status: </span><span id="status"></span></td>
-            </tr>
-          </tobdy>
-        <table>
-      </section>
-      <hr/>
-      <section>
-        <h2>Send Ether</h2>
-        <table>
-          <thead>
-            <tr>
-              <td>Send To:</td>
-              <td>Amount:</td>
-              <td></td>
-            </tr>
-          </thead>
-          <tobdy>
-            <tr>
-              <td><input type="text" id="sendTo" size="40" placeholder="0x1234567891234567"/></td>
-              <td><input type="text" id="amount" size="10" placeholder="5"/></td>
-              <td><button class="btn btn-sm btn-success" onclick="sendEther()">Send Ether</button></td>
-            </tr>
-            <tr>
-              <td><span>Transaction Hash(ID): </span><span id="txHashSentEth"></span></td>
-            </tr>
-          </tobdy>
-        </table>
-      </section>
-    </main>
-    <script>
-      // DOM Selector Shortcut for this example
-      let $ = (selector) => document.querySelector(selector)
-
-      // App Name reading for this example
-      // let dappName = $('title').textContent;
-      // $('#dappName').textContent = dappName;
-
-      ////////////////////////////////////////////////
-
-      // App Setup
-      let connect = new uportconnect.Connect('Uport Example')
-      let web3 = connect.getWeb3()
-
-      function loginBtnClick () {
-        const connectResponse = connect.requestCredentials().then(credentials => {
-          console.log(credentials)
-          $('#addr').textContent = credentials.address
-          $('#name').textContent = credentials.name
-        }, console.err)
-      }
-
-
-      ////////////////////////////////////////////////
-
-      // This contract ABI has already been generated for you
-      // Generating contract ABI's is out of scope of this document
-      // See Truffle for help with contract compiling
-      // https://truffle.readthedocs.io/en/latest/
-
-
-
-      const updateStatus = () => {
-        let statusContract = connect.contract(
-          [
-            {
-              "constant": false,
-              "inputs": [
-                {
-                  "name": "status",
-                  "type": "string"
-                }
-              ],
-              "name": "updateStatus",
-              "outputs": [],
-              "type": "function"
-            },
-            {
-              "constant": true,
-              "inputs": [
-                {
-                  "name": "addr",
-                  "type": "address"
-                }
-              ],
-              "name": "getStatus",
-              "outputs": [
-                {
-                  "name": "",
-                  "type": "string"
-                }
-              ],
-              "type": "function"
-            }
-          ]
-        )
-        let status = statusContract.at("0xB42E70a3c6dd57003f4bFe7B06E370d21CDA8087")
-        let statusMsg = $("#sts").value
-
-        status.updateStatus(statusMsg).then(tx => {
-          console.log(tx)
-          $("#txHashStatus").textContent = tx
-        })
-
-      }
-
-      ////////////////////////////////////////////////
-
-      // Send ether to a friend with Web3
-      // const sendEther = () => {
-      //   let toAddr = $('#sendTo').value
-      //   let value = parseFloat($('#amount').value) * 1.0e18
-      //   let gasPrice = 100000000000
-      //   let gas = 500000
-      //
-      //   web3.eth.sendTransaction(
-      //     {
-      //       to: toAddr,
-      //       value: value,
-      //       gasPrice: gasPrice,
-      //       gas: gas
-      //     },
-      //     (error, txhash) => {
-      //       if (error) { throw error }
-      //       $('#txHashSentEth').innerHTML = txhash
-      //     }
-      //   )
-      // }
-    </script>
-  </body>
-</html>
